@@ -1,24 +1,24 @@
 from dataclasses import dataclass
-from typing import Any, Awaitable, Callable, Dict, Optional, Type, Union
+from typing import Any, Awaitable, Callable, Type
 
 from aiogram import types
 
-from .filters import FormFilter, _FormFilter
+from .filters import FormFilter, InputTransformer
 
-Markup = Union[
-    types.ReplyKeyboardMarkup, types.InlineKeyboardMarkup, types.ReplyKeyboardRemove
-]
+Markup = (
+    types.ReplyKeyboardMarkup | types.InlineKeyboardMarkup | types.ReplyKeyboardRemove
+)
 
-EnterCallback = Callable[[int, int, Dict[str, Any]], Awaitable[types.Message]]
+EnterCallback = Callable[[int, int, dict[str, Any]], Awaitable[types.Message]]
 
 
 @dataclass(frozen=True)
 class FormFieldInfo:
-    enter_message_text: Optional[str]
-    error_message_text: Optional[str]
-    filter: Optional[FormFilter]
-    reply_markup: Optional[Markup]
-    enter_callback: Optional[EnterCallback]
+    enter_message_text: str | None
+    error_message_text: str | None
+    filter: FormFilter | None
+    reply_markup: Markup | None
+    enter_callback: EnterCallback | None
 
 
 @dataclass(frozen=True)
@@ -26,16 +26,16 @@ class FormFieldData:
     name: str
     type: Type
     info: FormFieldInfo
-    filter: _FormFilter
+    transformer: InputTransformer
 
 
 def FormField(
     *,
-    enter_message_text: Optional[str] = None,
-    filter: Optional[FormFilter] = None,
-    error_message_text: Optional[str] = None,
-    reply_markup: Optional[Markup] = None,
-    enter_callback: Optional[EnterCallback] = None
+    enter_message_text: str | None = None,
+    filter: FormFilter | None = None,
+    error_message_text: str | None = None,
+    reply_markup: Markup | None = None,
+    enter_callback: EnterCallback | None = None
 ) -> Any:
     if enter_message_text is None and enter_callback is None:
         raise ValueError("enter_message_text or enter_callback should be set")
