@@ -1,25 +1,16 @@
 import datetime
-from enum import IntEnum
-from typing import Any, Callable, NamedTuple, Union
+from dataclasses import dataclass
+from typing import Any, Callable, Coroutine, Protocol, Type
 
 from aiogram import F, types
 from aiogram.utils.magic_filter import MagicFilter
 
-FormFilter = Union[MagicFilter, Callable[..., Any]]
+from . import utils
 
+FormFilter = MagicFilter | Callable[..., Any]
+InputTransformResult = tuple[Any, bool]
 
-class _FormFilterType(IntEnum):
-    magic = 0
-    func = 1
-    coro = 2
-
-
-class _FormFilter(NamedTuple):
-    filter: FormFilter
-    filter_type: _FormFilterType
-
-
-DEFAULT_FORM_FILTERS = {
+DEFAULT_FORM_FILTERS: dict[Type, FormFilter] = {
     str: F.text,
     int: F.text.func(int),
     float: F.text.func(float),
@@ -36,3 +27,67 @@ DEFAULT_FORM_FILTERS = {
     types.Document: F.document.func(lambda document: document),
     types.Message: F.func(lambda m: m),
 }
+
+
+class InputTransformer(Protocol):
+    async def transform_input_message(
+        self, message: types.Message, data: dict[str, Any]
+    ) -> InputTransformResult: ...
+
+
+@dataclass(frozen=True)
+class MagicInputTransformer(InputTransformer):
+    filter: MagicFilter
+
+    async def transform_input_message(
+        self, message: types.Message, data: dict[str, Any]
+    ) -> InputTransformResult:
+        filter_result = self.filter.resolve(message)
+
+        if filter_result is None or filter_result is False:
+            return filter_result, False
+
+        return filter_result, True
+
+
+@dataclass(frozen=True)
+class SyncInputTransformer(InputTransformer):
+    filter: Callable[..., Any]
+
+    async def transform_input_message(
+        self, message: types.Message, data: dict[str, Any]
+    ) -> InputTransformResult:
+        prepared_field_filter = utils.prepare_function(self.filter, message, **data)
+        filter_result = prepared_field_filter()
+
+        if filter_result is False:
+            return filter_result, False
+
+        return filter_result, True
+
+
+@dataclass(frozen=True)
+class AsyncInputTransformer(InputTransformer):
+    filter: Callable[..., Coroutine[Any, Any, Any]]
+
+    async def transform_input_message(
+        self, message: types.Message, data: dict[str, Any]
+    ) -> InputTransformResult:
+        prepared_field_filter = utils.prepare_function(self.filter, message, **data)
+        filter_result = await prepared_field_filter()
+
+        if filter_result is False:
+            return filter_result, False
+
+        return filter_result, True
+
+
+def get_form_filter_for_type(field_type: Type) -> FormFilter:
+    field_filter = DEFAULT_FORM_FILTERS.get(field_type)
+
+    if field_filter is None:
+        raise TypeError(
+            f"There is no default filter for type {field_type}. You should consider writing your own filter"  # noqa: E501
+        )
+
+    return field_filter
